Separate DLL export/import for test helpers dll, and fix typo.
The DLL export issue isn't a problem currently, since we don't explicitly export any symbols from the test dll.  This will be an issue in the future though.  Additionally, there was a typo in absl_internal_test_dll_contains that caused abseil_test_dll to be empty (and therefore not create a lib file).

PiperOrigin-RevId: 504555797
Change-Id: Ic7b50bcbe704f7c8fd44028071abcf5d6babb2cf